requirements.txt
fastapi
uvicorn[standard]
pillow
matplotlib
openai
python-multipart
